Update: CraftCanvas
유닛 배치 및 아이템 생성

+++ Assets/EvolveThisMatch/Script/Lobby/Craft/DepartmentTemplate.cs
@@ -52,13 +52,13 @@ namespace EvolveThisMatch.Lobby
     public class CraftItemData
     {
         [SerializeField, Label("제작품")] private ObscuredIntVariable _variable;
-        [SerializeField, Label("제작 시간(초)")] private float _craftTime;
-        [SerializeField, Label("무게(g)")] private float _weight;
+        [SerializeField, Label("제작 시간(초)")] private int _craftTime;
+        [SerializeField, Label("무게(g)")] private int _weight;
         [SerializeField] private RequiredItem[] _requiredItems;
 
         internal ObscuredIntVariable variable => _variable;
-        internal float craftTime => _craftTime;
-        internal float weight => _weight;
+        internal int craftTime => _craftTime;
+        internal int weight => _weight;
         internal IReadOnlyList<RequiredItem> requiredItems => _requiredItems;
 
         [Serializable]
